# GestorBarberia.Api/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GestorBarberia.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Estilosdecorte/Estilos/{id}": {
      "get": {
        "tags": [
          "Estilosdecorte"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Estilosdecorte/EstilosByBarberiaId/{id}": {
      "get": {
        "tags": [
          "Estilosdecorte"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Estilosdecorte/EstilosByClienteBarberiaId/{id}": {
      "get": {
        "tags": [
          "Estilosdecorte"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Estilosdecorte/Save": {
      "post": {
        "tags": [
          "Estilosdecorte"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EstilosdecorteAddDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Estilosdecorte/Update": {
      "put": {
        "tags": [
          "Estilosdecorte"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EstilosdecorteUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Estilosdecorte/Delete": {
      "delete": {
        "tags": [
          "Estilosdecorte"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/EstilosdecorteRemoveDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Estilosdecorte/DeleteByAdmin/{estilioId}/{adminId}": {
      "delete": {
        "tags": [
          "Estilosdecorte"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "estilioId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "adminId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reportes/GetReportes/{barberiaId}/{fechaInicio}/{fechaFin}": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "barberiaId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "fechaInicio",
            "required": true,
            "type": "string",
            "format": "date-time"
          },
          {
            "in": "path",
            "name": "fechaFin",
            "required": true,
            "type": "string",
            "format": "date-time"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reportes/GetIngresos/{barberiaId}": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "barberiaId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reportes/GetTotalClientes/{barberiaId}": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "barberiaId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Reportes/GetTotalBarberos/{barberiaId}": {
      "get": {
        "tags": [
          "Reportes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "barberiaId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
...
    "EstilosdecorteAddDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "estiloId": {
          "format": "int32",
          "type": "integer"
        },
        "barberiaId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "imgestilo": {
          "type": "string"
        }
      }
    },
    "EstilosdecorteRemoveDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "estiloId": {
          "format": "int32",
          "type": "integer"
        },
        "barberiaId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "imgestilo": {
          "type": "string"
        }
      }
    },
    "EstilosdecorteUpdateDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "estiloId": {
          "format": "int32",
          "type": "integer"
        },
        "barberiaId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "imgestilo": {
          "type": "string"
        }
      }
    },
...
  }
}