# GestorBarberia.Application/Dtos/EstilosdecorteDto/EstilosdecorteRemoveDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace GestorBarberia.Application.Dtos.EstilodecorteDto
{
    public class EstilosdecorteRemoveDto : EstilosdecorteDto
    {
    }
}